refactor(network): consolidate network interface retrieval into sysinfo plugin
- Move network interface formatting and retrieval logic from main backend to `tauri-plugin-clash-verge-sysinfo`.
- Unify redundant `get_network_interfaces` and `get_network_interfaces_info` commands into a single optimized implementation.
- Remove unnecessary `AsyncHandler` overhead for lightweight system queries.
- Fix MAC address serialization by explicitly handling `None` for `00:00:00:00:00:00` fallbacks.
- Update frontend hooks and components to consume the new detailed interface.
