+++ crates/tauri-plugin-clash-verge-sysinfo/Cargo.toml
@@ -11,6 +11,8 @@ parking_lot = { workspace = true }
 # sysinfo 0.38.2 conflicts with dark-light
 # see https://github.com/GuillaumeGomez/sysinfo/issues/1623
 sysinfo = { version = "0.39.2", features = ["network", "system"] }
+serde = { workspace = true }
+serde_json = { workspace = true }
 
 [target.'cfg(not(windows))'.dependencies]
 libc = "0.2.186"
