+++ src-tauri/src/lib.rs
@@ -144,7 +144,6 @@ mod app_init {
             cmd::open_app_log,
             cmd::open_core_log,
             cmd::get_portable_flag,
-            cmd::get_network_interfaces,
             cmd::get_system_hostname,
             cmd::restart_app,
             cmd::start_core,
