+++ src/services/cmds.ts
@@ -431,10 +431,6 @@ export async function downloadIconCache(url: string, name: string) {
   return invoke<string>('download_icon_cache', { url, name })
 }
 
-export async function getNetworkInterfaces() {
-  return invoke<string[]>('get_network_interfaces')
-}
-
 export async function getSystemHostname() {
   return invoke<string>('get_system_hostname')
 }
